+++ pkg/webui/console/store/selectors/organizations.js
@@ -26,6 +26,7 @@ import {
   createEventsSelector,
   createEventsErrorSelector,
   createEventsStatusSelector,
+  createEventsInterruptedSelector,
 } from './events'
 import { createFetchingSelector } from './fetching'
 import { createErrorSelector } from './error'
@@ -66,3 +67,4 @@ export const selectOrganizationRightsFetching = createFetchingSelector(GET_ORGS_
 export const selectOrganizationEvents = createEventsSelector(ENTITY)
 export const selectOrganizationEventsError = createEventsErrorSelector(ENTITY)
 export const selectOrganizationEventsStatus = createEventsStatusSelector(ENTITY)
+export const selectOrganizationEventsInterrupted = createEventsInterruptedSelector(ENTITY)